refactor: Trim out unnecessary functions
Remove the defineViewport and setVisibilityLists functions.  The code
functions just the same when using the $sections variable in the
controlFade function in place of the $sectionsOutViewport variable.
Code still not functioning as ultimately desired, but this change
should help make it easier to make the right changes and reach the
desired final functionality.